+++ app-frontend/src/app/components/pages/project/index.js
@@ -4,6 +4,8 @@ import autoInject from '_appRoot/autoInject';
 import layerPages from './layer';
 import projectPage from './project';
 import layersPage from './layers';
+import createAnalysisPage from './createAnalysis';
+import analysisPage from './analysis';
 import settingsPages from './settings';
 import analysesPages from './analyses';
 
@@ -47,6 +49,8 @@ export {
 export default [
     layersPage,
     projectPage,
+    createAnalysisPage,
+    analysisPage,
     ...layerPages,
     ...settingsPages,
     ...analysesPages