+++ app-frontend/src/app/components/projects/layerItem/index.js
@@ -10,7 +10,7 @@ class LayerItemController {
     $onInit() {
         const rx = /^#(?:[0-9a-f]{3}){1,2}$/i;
         const color = _.get(this.itemInfo, 'colorGroupHex');
-        if (color.match(rx)) {
+        if (color && color.match(rx)) {
             this.color = color;
         } else {
             this.color = 'white';
@@ -26,9 +26,9 @@ const component = {
         itemInfo: '<',
         itemActions: '<',
         selected: '<',
-        onSelect: '&',
+        onSelect: '&?',
         visible: '<',
-        onHide: '&',
+        onHide: '&?',
         isAnalysis: '<?',
         isExport: '<?',
         onDownloadExport: '&?'